pyproject.toml
[project]
name = "minimal_agent"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "luyanfcp", email = "luyanfcp@foxmail.com"},
]
dependencies = ["pydantic>=2.11.0", "requests>=2.32.3", "openai>=1.70.0", "rich>=14.0.0", "opentelemetry-api>=1.31.1", "opentelemetry-sdk>=1.31.1", "opentelemetry-exporter-otlp>=1.31.1", "dashscope>=1.23.0", "bs4>=0.0.2", "markdownify>=1.1.0"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[dependency-groups]
dev = [
    "pyright>=1.1.398",
    "ruff>=0.11.2",
    "black>=25.1.0",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
]
